refactor: extract education plan composable
